Extract fatigue compliance boundary
