+++ tests/test_routes.py
@@ -23,6 +23,7 @@ if os.path.exists(TEST_DB_PATH):
 os.environ["DATABASE_URL"] = f"sqlite:///{TEST_DB_PATH}"
 
 import app
+import fatigue_compliance
 import fatigue_engine
 from app import (
     AnnotationAudit,
@@ -1481,6 +1482,11 @@ def test_d24_requires_a_complete_observation_window(client):
             observation_start=start - timedelta(days=30),
         )
     assert app._analyze_segments is fatigue_engine._analyze_segments
+    assert app._compliance_month is fatigue_compliance.compliance_month
+    assert (
+        app._fatigue_rule_config
+        == app._fatigue_rule_config_service.load
+    )
     assert not any(
         message.startswith("D24:")
         for messages in without_history.values()
